### 2018-02-07 #195 @willpearse
submission: suppdata
_Downloads supplementary materials from published papers using their DOIs as a reference. This facilitates reproducible analyses._

```
Package: suppdata
Title: Downloading Supplementary Data from Published Manuscripts
Version: 0.9-0
Author: William D. Pearse, Scott Chamberlain
Maintainer: William D. Pearse <will.pearse@gmail.com>
Description: Downloads data supplementary materials from manuscripts,
    using papers' DOIs as references. Includes some code to download
    from other Internet APIs (e.g., Xeno-Canto, EPMC).
LazyLoad: yes
    knitr (>= 1.6),
    fulltext(>= 0.1.4.9000)
    xml2 (>= 1.2.0),
    jsonlite (>= 1.5),
    RCurl (>= 1.95-4.10),
    rcrossref (>= 0.8.0)
```

-   URL for the package: [https://github.com/willpearse/suppdata](https://github.com/willpearse/suppdata)

  - _data retrieval_: because it downloads data from online scientific publications.
  - _reproducibility_: because by downloading raw data, it facilitates reproducible analysis scripts

  - _Audience_: anyone (re-)analysing published scientific data
  - _Application_: Conducting reproducible analyses, database creation scripts

_Originally, this code was part of `fulltext`, but has now been split out. That is the only package I'm aware of that does this._

  @sckott

  - _This package does download data from journal publishers such as Wiley. I have heard stories/rumours of them getting very unhappy about people downloading articles from their websites at high speed. I'm not aware that this package violates their terms/conditions, but I want to flag this here. This code has already been part of an ROpenSci package (`fulltext`) without any problems, so I can't see that this is an issue._

    - [x] The package is deposited in a long-term repository with the DOI: 0.5281/zenodo.1168713
	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

  _Yes it does, and all this code was, previously, part of the ROpenSci package `fulltext`._
