### 2018-02-07 #195 @noamross
Thanks @willpearse for your submission (and creation of this package at our behest!).  Things look good, but I haven't been able to run all of our tests because of this build error, which I get both locally and on r-hub:

```
* checking for file ‘suppdata/DESCRIPTION’ ... OK
* preparing ‘suppdata’:
* creating vignettes ... ERROR
Error in texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet,  :
  Running 'texi2dvi' on 'suppdata-intro.tex' failed.
LaTeX errors:
! LaTeX Error: File `ae.sty' not found.

Type X to quit or <RETURN> to proceed,
or enter new name. (Default extension: sty)

! Emergency stop.
<read *>

l.30 \ifthenelse
                {\boolean{Sweave@inconsolata}}{%^^M
Calls: <Anonymous> -> texi2pdf -> texi2dvi
Execution halted
```

It's odd that this doesn't show up on Travis.  Anything non-standard in your vignette Rnw?
