### 2018-02-07 #195 @willpearse
No worries, glad to help.

I've honestly no idea why this is happening (because, as you say, it
doesn't trigger on Travis or on my machine), but from a quick bit of
Googling I think it may be to do with your LaTeX install? Try:

devtools::install_github(c('yihui/tinytex', 'rstudio/rmarkdown'))
# first uninstall your existing LaTeX distribution, and then
tinytex::install_tinytex()

...following from https://github.com/rstudio/rmarkdown/issues/359
