### 2018-02-07 #195 @noamross
So I tried building the vignette in a couple of fresh Docker images: `rocker/verse` (which has LaTeX pre-installed)`, and `rocker/tidyverse`, which does not, using *tinytex* to install TeX.  Both gave me the same error, which means that somehow your vignette is doing something that is not in either of those standard environments, but *is* on Travis and your machine.  Oddly the `ae` package, which should have the `ae.sty` file, _is_ installed with tinytex.
