### 2018-02-07 #195 @willpearse
Have you followed the instructions I gave travis? Like `fulltext`, this
needs `libpoppler-cpp-dev` installed. (
https://github.com/willpearse/suppdata/blob/master/.travis.yml).
