### 2018-02-07 #195 @noamross
OK, I've solved this by installing `texinfo` into my Docker container, but that doesn't seem like a long-run solution.  I'll ask reviewers to see if they have issues and we'll return to it if they do. I might suggest a switch to markdown if it keeps up.

### Editor checks:

- [ ] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.  ***Please add code coverage to your CI and a badge in your readme (see `usethis::use_coverage`

---

#### Editor comments

Here's the `goodpractice::gp()` output.  Nothing to prevent moving to review but these should be straightforward fixes.  I will go ahead and find reviewers.

```
 ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite
    often. A build date will be added to the package when you perform `R CMD build` on
    it.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    R/journals.R:9:1
    R/journals.R:18:1
    R/journals.R:24:1
    R/journals.R:33:1
    R/journals.R:36:1
    ... and 47 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list,
    depending on the input data. Consider using vapply() instead.

    R/suppdata.R:126:17
    R/suppdata.R:135:17

  ✖ fix this R CMD check NOTE: Namespace in Imports field not imported from:
    ‘RCurl’ All declared Imports should be used.
───────────────────────────────────────────
---

Reviewers:
Due date:
