### 2018-02-08 #195 @willpearse
Thanks for your effort on this, and I'm sorry it gave you so much trouble
to install.

I'll wait to hear back from the reviewers, and then I'll make all these
changes and whatever they request in one fell swoop.

Thanks again!
