### 2018-02-08 #195 @noamross
Oh, one other thing, I also get the following error while testing (*goodpractice* runs as CRAN so it doesn't show up):
```
test-all.r:52: warning: suppdata fails well
404: Resource not found. - (nonsense)
```
I believe this is due to a breaking API change in **testthat** and you  want `expect_warning()`.
