### 2018-03-09 #195 @rossmounce
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

There is a clear need for an R package like this, to download suppdata based upon a publication identifier such as a DOI.

As it currently stands this package is too thin in my opinion - it needs more development before being accepted in JOSS. It barely goes beyond what is already possible with `utils::download.file` https://github.com/willpearse/suppdata/issues/20#issuecomment-371820342 The statement of need needs to be clearer and more explicit in terms of subject disciplines covered and scale of analyses intended.

AFAIK there is no example given of how to set the download directory parameter correctly https://github.com/willpearse/suppdata/issues/14

The README.md file is woefully insufficient on the intended scope/usage of the package and in informational content in general, it needs a whole lot more than just one overly simplistic example https://github.com/willpearse/suppdata/issues/11 https://github.com/willpearse/suppdata/issues/18 https://github.com/willpearse/suppdata/issues/20#issuecomment-371820342 . I had a look at other JOSS packages just to be sure about this e.g. https://github.com/ropensci/drake and https://github.com/Molmed/checkQC and yes both README files on github at those two projects are far more helpfully detailed. AFAIK github is a valid first point of entry for a user wanting to use the software.

On a higher-level, some thought needs to be given as to what and who (discipline-wise?) this package is for and if at the moment it could reasonably fulfil a real use-case e.g. comparing supplementary data between _PeerJ_ and _Nature_ https://github.com/willpearse/suppdata/issues/19  Even if one scoped the intetion of this package down to just 'ecology journals' it still couldn't fulfil an ecological meta-analysis use-case. I hate journal rankings and this is in no way an endorsement of journal rankings but just for the sake of argument, if one takes a look at what are considered by one citation-based metric (SJR) the 'top 50' ecology journals, I'd say this package covers suppdata at less than half http://www.scimagojr.com/journalrank.php?category=2303

I humbly suggest given the scale of and diversity of suppdata at academic journals that more thought is given to documenting how external community contributors can get involved with the project to increase the diversity and coverage of journals. I have given lots of concrete examples as a constructive starting point https://github.com/willpearse/suppdata/issues/2 https://github.com/willpearse/suppdata/issues/18

FYI linking data with literature is a hot topic at the moment. Have a look at http://www.scholix.org/ for instance. This might provide a nice pathway in future.

As I said in issue 18 https://github.com/willpearse/suppdata/issues/18 if Will would like, I'd love to help with increasing journal coverage if I can, but I'm not too familar with implementing regular expressions in R - I'm not sure how best to contribute. This package has a lot of potential though.
