### 2018-03-10 #195 @noamross
Thank you for your thoughtful review, @rossmounce!  @willpearse, you can wait until @sarahsupp's review is in to respond, and I'll provide an editor's summary as well.
