### 2018-03-12 #195 @sarahsupp
>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

2.5

---

As someone who has done many studies using published supplementary data, I think this package is a great idea! While it seems to meet a clear need, it is lacking Statement of Need, with a clear description of who the intended user is, and how the package will help automate their analytical process in R. Some of this info is presented in the paper.md file, but not in the README.md.

For example, while it seems like many major publishers are supported (i.e., Wiley, Science, Biorxiv), with decent representation from biological/ecological sources, it is unclear to me if this package would be most beneficial only to those fields, or if it has strong enough representation for use in other STEM research areas, or quantitative research in the social sciences. It would be excellent to see something like this serve a wide diversity of fields. Even within ecology, I could see myself using this package for a relatively small analysis that used 1 or a few datasets from supported journals, with a streamlined code from raw data to analysis; but for larger meta-analyses, only a portion of the datasets I'd need access to look like they'd be supported, as the package currently stands. Do the authors plan to add more publishers and data repositories to the package, or are there some limitations that prevent doing so?

In the R help files for help(suppdata), I was not sure what the authors meant when they said "I'm aware that there isn't perfect overlap between these publishers and the rest of the package; I plan to correct this in the near future."

The help file was very useful, and since some of the publishers require different ways of calling the data, the examples were much needed and useful. There are a few small typos in the help file (SCott, FigShare).

The package conforms to most of the rOpenSci packaging guidelines, but is missing some of the requested information (see check boxes above).

I could not run the vignette successfully locally, which I tried to do using RSweave, Compile PDF. When I tried to compile the pdf, I had errors associated with:
* Line 38  Citation 'crabs' on page 1 undefined on input line 38.
* Line 105  Font T1/cmr/m/n/12=ecrm1200 at 12.0pt not loadable: Metric (TFM) file not found.

My disclaimer is that I do not have a lot of experience with .Rnw files, so perhaps it is a user (me) error.

Overall, I think the package has a lot of great potential, but perhaps needs a bit more documentation and if it is feasible, greater connection to supplemental data hosted across additional publishers and repositories.
