### 2018-03-12 #195 @rossmounce
Thanks @sarahsupp , I also had problems with the vignette but wasn't sure if it was just me and my particular setup, so I didn't say anything about it.
