### 2018-03-13 #195 @noamross
Thanks for your review, @sarahsupp!

Let me chime on on the biggest issue that was highlighted in both Sarah and Ross's reviews: that of journal coverage and the ability to collect supplemental data en masse.  This is a "long tail" problem, of which we are quite familiar at rOpenSci.  Ross's suggestion that one way to deal with this is facilitate community contributions is a good one: provide good documentation in a vignette or CONTRIBUTING file of how to provide a publisher plugin with appropriate tests, along with up-front documentation of the current journals covered and the next one to prioritize.   As an rOpenSci package we'd be happy to facilitate some this maintenance - reviewing pull requests, using our fora as a way to encourage others to contribute, etc.,  So the long-term burden won't be all on @willpearse :)

I note a project both @rossmounce and I were involved with, [**quickscrape**](https://github.com/ContentMine/quickscrape), went down this road, going so far to to write a JSON spec for [web site scrapers for different publishers](https://github.com/ContentMine/journal-scrapers/tree/master/scrapers). That repo may be a good resource for some additional patterns for **suppinfo**.

Thanks for the mention of Scholix - rOpenSci may build something specific for it in the future (pinging @sckott, who may already know about it). We would, of course, prefer if a central standard and clearinghouse allowed us to get most Supplementary Data.  I believe our **fulltext** package is able to find full-text articles through CrossRef for some publishers, but still needs **suppdata**-like handlers for many that don't enable full-text access this way.
