### 2018-03-15 #195 @sckott
It is indeed a really hard problem to programatically sort out links for articles, and even harder for supp. materials. In `fulltext`, we take a variety of approaches, starting from the highest level crossref, then if not found there, moving down to various other tools such as publisher specific code in the pkg, or using an API https://ftdoi.org/ I made to sort out URL patterns per publisher. That API - and as far as I can tell quickscrape as well - don't have any logic for supp. materials. A community contributed set of logic does seem best as a standalone thing - to be used in any programming language and via an API like ftdoi.org or any other.

What % of supp. materials files are now in Figshare/Dryad/other repositories?  How useful is it to ping those APIs or to ping Datacite API?
