### 2018-03-16 #195 @rossmounce
@sckott  "What % of supp. materials files are now in Figshare/Dryad/other repositories?"

That's a really good question I'd also love to have an answer to. To re-state it slightly
What proportion of SI is at third-party repositories versus journal-hosted

The confusing bit is that for some journals e.g. PLOS the SI is available at _both_ figshare _and_ the journal platform itself.

I have a feeling that for say articles published in 2014 the % of SI that is journal hosted is greater than the percentage that is third-party hosted. But maybe that (horrid) situation is improving somewhat. Articles published in 2017 might be expected to have a higher proportion of SI at third-party repositories?  There is of course disciplinary variation and journal to journal variation. Depends highly upon one's scoping!
