### 2018-04-08 #195 @willpearse
I thank the reviewers and editor for their comments and suggested improvements to `suppdata`. I think that, as a result of their suggestions, the package is in much better shape. Thank you!

Below I respond to what the editor highlighted as the over-arching concern, and after that I respond to more specific issues raised by the reviewers and the editor.

## Supporting more publishers and the package being "too thin"

I agree with the reviewers and editor that `suppdata` needs more publisher wrappers. To address this need going forward, I have created a detailed wiki-page for the package that describes how to go about adding new wrappers, in the hope that the community can help with this. This is referenced in the readme file, and can be found here: https://github.com/willpearse/suppdata/wiki.

Dr Supp raised a concern that the current coverage might be insufficient to facilitate a large meta-analysis, and Dr Mounce raised similar concerns about the top-50 journals in Ecology. While I, of course, agree that `suppdata` is not appropriate for everything, I have written two packages - [NATDB](https://github.com/willpearse/natdb) and [NACDB](https://github.com/pearselab/nacdb) - that, combined, load of the order of 130 studies' datasets using `suppdata`. Thus I do agree that `suppdata` isn't perfect, but it is useful (at least to me and my collaborators) right now.

Dr Mounce raises the concern that `suppdata` doesn't cover as many publishers as ContentMine. He says ContentMine covers ~30 publishers; `suppdata` covers ~10 (https://github.com/willpearse/suppdata/issues/18). ContentMine is a company with multiple employees (at the time of writing, they are recruiting to fill two positions); I am a single developer, and so to be honest I'm quite pleased that `suppdata` covers even a third of the number of publishers that ContentMine does. For clarity, I add that my understanding is that ContentMine is similar in scope and aim to what the package `fulltext` carries out, not what `suppdata` does, but I would be grateful of being corrected if I am wrong.

I do agree that more work is needed on `suppdata`; my hope is the package can get that help as a member of the ROpenSci ecosystem. I add that I was asked by ROpenSci to submit this code as a separate package, so I presume that is also the hope of others at ROpenSci.

## More specific comments

* I have added more documentation to the package on its GitHub page. I would be grateful if the reviewers could let me know if this is sufficient for them, and if not where they would like more documentation. I don't think this should be a difficult package to use, and that's why I've tried to keep the documentation short, sharp, and sweet. I could easily have failed in my attempt, and so, again, any suggestions would be gratefully received!
* Vignette build problems.
 I have corrected the "crabs" citation error, but it is not immediately obvious to me what the error Dr Supp found on line 105 relates to. There is no line 105 in the vignette file, so the error is extremely opaque to me. All I know is that the vignette builds on Travis, which implies to me that whatever is stopping the reviewers from building the package is specific to their system. Perhaps there are two things the reviewers could try doing:
  * Make sure they have LaTeX installed on their system (see https://support.rstudio.com/hc/en-us/articles/200532257-Customizing-LaTeX-Options and links)
  * If they're on Linux, try running `sudo apt install libpoppler-cpp-dev`, which is a PDF rendering library that they might be missing.
  * If neither of those steps works, if the reviewers could let me know what operating system they're using I'll look into it further. Regardless, I have added a PDF to the GitHub repo now, so it not necessary to build the vignette from source.
* Lack of examples for user-facing functions. All user-facing functions have examples (there is only one; `suppdata`). They are wrapped in "dontrun" blocks to avoid having all the CRAN mirrors spamming servers with download requests for the same files. This is described in the documentation, but if the user runs the examples themselves they do run. CRAN prefers for examples not to take longer than five seconds anyway; if I were to remove them from this "dontrun" block then I would likely get into trouble with CRAN. If the reviewers are unhappy with this, I'm happy to make whatever changes they suggest as long as they won't get me into trouble when I submit to CRAN.
* I have changed the version number to 1.0-0
* The only complaint from goodpractice::gp is that only 93% of the package is unit tested.
* I have added a review badge to the README, along with a covr badge
* I have added a statement of need to the README ("a more detailed set of motivations for `suppdata`")
* The phrase "I'm aware that there isn't perfect overlap between these publishers and the rest of the package; I plan to correct this in
 the near future" was a hang-over from when this code was part of `fulltext`; thank you for finding this, and I have now removed it from the helpfile.
