### 2018-04-13 #195 @noamross
Thanks for your response, @willpearse!  @rossmounce and @sarahsupp, please let us know if Will's response addresses your comments.

A quick suggestion, Will: I suggest moving content from the GH wiki to a CONTRIBUTING.md file and README/other documentation, which will travel more transparently with clones and forks of the repo.
