### 2018-04-16 #195 @sarahsupp
Hi @willpearse. I think the new wiki pages will be really useful in encouraging folks to help contribute and in giving more description into why and how to use the suppdata package. Nice addition!

For the vignette build issue, I was working on Linux, and I think it's entirely possible that the PDF rendering library was missing. I can check later, and I'll try to get back to you if I figure out what the problem was. The PDF should help in case others have this problem too, or aren't used to the process of building the vignette from source.

While I mentioned previously that the number of publishers currently limits the usefulness of the package for large meta-analyses, I do think it is a great start and can already be useful for many projects, and I wouldn't consider this a major concern for moving forward. In general, I think that suppdata is looking good, and I'm excited to see how this grows as more can contribute to it and expand the publisher base from which data can be accessed.
