### 2018-04-25 #195 @noamross
@rossmounce Can you let us know if Will's update addresses your comments?
