### 2018-04-26 #195 @rossmounce
Thanks for the ping. I think Will's updates mostly address my comments.

It's a great start and I look forward to seeing the usefulness of this package grow over time with more scraper methods to cover a wider range of publishers.
