### 2018-04-29 #195 @noamross
Approved!  Thank you @willpearse, @sarahsupp, and @rossmounce.  To-dos:

To-dos:
- [ ] Update badges, CI and coverage links, and URLs to point to the new repo address.
- If you would like to, and if your reviewers agree, you may choose to credit their reviews in the DESCRIPTION file by adding them as `'rev'`-type contributors ([example](https://github.com/ropensci/rrricanes/blob/master/DESCRIPTION#L11-L12))
- [ ] Make a new Zenodo archive from the updated repo
- [ ] Submit to JOSS and note this review thread

Welcome aboard! I think that it would be great to have a blog post about **suppdata**, especially as a mechanism to get more contributors to participate and add more publisher coverage.  If you are interested in writing one, let us know and, @stefaniebutland will be in touch about content and timing.
