### 2018-04-30 #195 @sarahsupp
Sure, I would be OK with being acknowledged as a rev-type contributor.
Thanks! I look forward to seeing how the package grows, and hopefully
getting a chance to play with it!

On Mon, Apr 30, 2018 at 4:21 PM, Will Pearse <notifications@github.com>
wrote:

> @noamross <https://github.com/noamross> awkwardly, I just went to
> settings, clicked 'transfer ownsership', gave it to ropensci (or so I
> thought), and was told that I didn't have permission to create a repo
> there. Have I clicked the wrong button? Sorry!...
>
