### 2018-05-01 #195 @rossmounce
Yes I'm OK with being acknowledged as a rev-type contributor. Thanks for asking.
