### 2018-05-02 #195 @noamross
@stefaniebutland Will has expressed interest in writing a blog post for **suppdata**.  I've pointed him to the guidance in the **roweb2** repo.  Let him know what the schedule is!
