### 2018-05-14 #195 @stefaniebutland
@willpearse I apologize for my delay in responding. I am temporarily putting most blog post reviews and scheduling on hold until after our late-May unconference. In the meantime, if it's helpful to you, you could proceed with a draft while it's relatively fresh in your mind. There is so much value to readers in these posts by package authors!

I anticipate resuming non-unconference blog post reviews in mid-June for publishing in July and August
