### 2018-05-16 #195 @willpearse
Thanks for this. I'm happy to leave this for a while; please do get back in touch with me closer to this being published. I am still adding things to the package, so I don't really want to spend time writing a blog asking for help developing features when it's reasonably likely I'll have implemented a lot of them by then. I'm still very happy to write the blog post, but I want to make sure it's optimally useful. Thanks for your help!
