### 2018-05-16 #195 @willpearse
Great, thanks very much!

On Wed, 16 May 2018 at 10:13 Stefanie Butland <notifications@github.com>
wrote:

> Thank you @willpearse <https://github.com/willpearse>. Your strategy
> makes perfect sense. I'll get back in touch in June
>
