### 2018-07-18 #195 @stefaniebutland
@willpearse Are you still interested in contributing a blog post? Still adding to the package or ready to write? I'm now starting to line up posts for publication in late August, September.
