### 2018-07-18 #195 @willpearse
Thanks for the nudge. I'm absolutely swamped right now, to be honest, so
I'd rather not write something for a month or so.

On Tue, 17 Jul 2018 at 23:50 Stefanie Butland <notifications@github.com>
wrote:

> @willpearse <https://github.com/willpearse> Are you still interested in
> contributing a blog post? Still adding to the package or ready to write?
> I'm now starting to line up posts for publication in late August, September.
>
