### 2018-08-21 #195 @stefaniebutland
@willpearse how about now? 😀 Would love to have a post about `suppdata` but I understand this goes beyond the time already committed to the review process. I have blog post publication dates available Sept 18, Oct 2, 9, 16. If you're still interested, let me know your preferred date and mark your calendar to submit a draft a week in advance.
